// src/app/models/assignment.model.ts
export class AssignmentModel {
    submitted: boolean;
    name: string;
    lastName: string;
    phoneNumber: string;
    vehicleType: string;
    hullSize: string;
    boatModel: string;
    engineType: string;
    registrationNumber: string;
    latitude: number;
    longitude: number;
    workerId : string;
    clientName: string;
    workerName: string;
    status: string;
    creationDateTime: Date;
}